+++ skribenten-backend/src/test/kotlin/no/nav/pensjon/brev/skribenten/auth/AuthorizeAnsattSakTilgangTest.kt
@@ -29,6 +29,8 @@ import java.time.LocalDate
 import java.time.Month
 import kotlin.test.Test
 import kotlin.test.assertEquals
+import kotlin.test.assertFalse
+import kotlin.test.assertTrue
 
 private const val NAVIdent = "månedens ansatt"
 private val testSakEnhet = NAVEnhet("en veldig bra enhet", "NAVs beste!")
@@ -300,6 +302,20 @@ class AuthorizeAnsattSakTilgangTest {
 
         val response = client.get("/sak/${sakVikafossen.saksId}")
         assertEquals(HttpStatusCode.InternalServerError, response.status)
+    }
+
+    @Test
+    fun `should return true if one of the penSakEnheter ids matches one of the navAnsattEnheter ids`() {
+        val navAnsattEnheter = listOf(NAVEnhet("1", "Enhet1"), NAVEnhet("2", "Enhet2"))
+        val penSakEnheter = listOf("2", "3")
+        assertTrue(harTilgangTilSakSinEnhet(navAnsattEnheter, penSakEnheter))
+    }
+
+    @Test
+    fun `should return false if none of the penSakEnheter ids matches navAnsattEnheter ids`() {
+        val navAnsattEnheter = listOf(NAVEnhet("1", "Enhet1"), NAVEnhet("2", "Enhet2"))
+        val penSakEnheter = listOf("3", "4")
+        assertFalse(harTilgangTilSakSinEnhet(navAnsattEnheter, penSakEnheter))
     }
 
     private fun successResponse(saksId: String) =
